feat: Add script to dynamically select docker or podman
This script is created to check for the container engine the user
is running. After checking for the engine it also checks the cgroup API
they are using. If they are using cgroup v2 and they have docker installed,
the script will automatically switch to podman. If the user has cgroup v1 enabled,
and the docker engine is installed, the script will use docker. Else it will default to
podman. If neither is installed or can be used, it will exit with non 0 code. Podman can
be used with both cgroup API versions as it is daemon-less and run rootless by default.
This script is necessary for developers that aren't running  the docker engine or that
their OS doesn't support the docker engine like some Linux distros (eg Fedora).